[update] optimize page youtube and page render video, completed page welcome
