+++ sources/dev/public/material/css/material-dashboard.css
@@ -15307,7 +15307,7 @@ fieldset[disabled] .form-group.is-focused .togglebutton label {
 
 .sidebar .sidebar-wrapper {
   position: relative;
-  height: calc(100vh - 75px);
+  height: calc(100vh - 120px);
   overflow: auto;
   width: 260px;
   z-index: 4;
