+++ sources/dev/public/styles.css
@@ -3,7 +3,6 @@
  */
 body {
 	padding-bottom: 2rem;
-	padding-top: 4rem;
 }
 .row {
 	margin-bottom: 1rem;
