+++ sources/dev/resources/sass/material-dashboard/_sidebar-and-main-panel.scss
@@ -32,7 +32,7 @@
 
     .sidebar-wrapper{
         position: relative;
-        height: calc(100vh - 75px);
+        height: calc(100vh - 120px);
         overflow: auto;
         width: 260px;
         z-index: 4;
