+++ sources/dev/resources/views/auth/verify.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app', ['class' => 'off-canvas-sidebar', 'activePage' => 'home', 'title' => __('Material Dashboard')])
+@extends('layouts.app', ['class' => 'off-canvas-sidebar', 'activePage' => 'home', 'title' => __('HQ Tool')])
 
 @section('content')
 <div class="container" style="height: auto;">
@@ -6,24 +6,24 @@
       <div class="col-lg-7 col-md-8">
           <div class="card card-login card-hidden mb-3">
             <div class="card-header card-header-primary text-center">
-              <p class="card-title"><strong>{{ __('Verify Your Email Address') }}</strong></p>
+              <p class="card-title"><strong>{{ __('Xác nhận địa chỉ email của bạn') }}</strong></p>
             </div>
             <div class="card-body">
               <p class="card-description text-center"></p>
               <p>
                 @if (session('resent'))
                     <div class="alert alert-success" role="alert">
-                        {{ __('A fresh verification link has been sent to your email address.') }}
+                        {{ __('Một link xác nhận đã được gửi vào địa chỉ email của bạn.') }}
                     </div>
                 @endif
                 
-                {{ __('Before proceeding, please check your email for a verification link.') }}
+                {{ __('Vui lòng kiểm tra email xác nhận link.') }}
                 
                 @if (Route::has('verification.resend'))
-                    {{ __('If you did not receive the email') }},  
+                    {{ __('Nếu bạn không nhận được email') }},  
                     <form class="d-inline" method="POST" action="{{ route('verification.resend') }}">
                         @csrf
-                        <button type="submit" class="btn btn-link p-0 m-0 align-baseline">{{ __('click here to request another') }}</button>.
+                        <button type="submit" class="btn btn-link p-0 m-0 align-baseline">{{ __('Nhấn vào đây để gửi một yêu cầu khác') }}</button>.
                     </form>
                 @endif
               </p>
