+++ sources/dev/resources/views/layouts/footers/guest.blade.php
@@ -1,35 +1,30 @@
 <footer class="footer">
     <div class="container">
-        <nav class="float-left">
-        <ul>
-            <li>
-            <a href="https://www.creative-tim.com">
-                {{ __('Creative Tim') }}
-            </a>
-            </li>
-            <li>
-            <a href="https://creative-tim.com/presentation">
-                {{ __('About Us') }}
-            </a>
-            </li>
-            <li>
-            <a href="http://blog.creative-tim.com">
-                {{ __('Blog') }}
-            </a>
-            </li>
-            <li>
-            <a href="https://www.creative-tim.com/license">
-                {{ __('Licenses') }}
-            </a>
-            </li>
-        </ul>
-        </nav>
-        <div class="copyright float-right">
-        &copy;
-        <script>
-            document.write(new Date().getFullYear())
-        </script>, made with <i class="material-icons">favorite</i> by
-        <a href="https://www.creative-tim.com" target="_blank">Creative Tim</a> and <a href="https://www.updivision.com" target="_blank">UPDIVISION</a> for a better web.
-        </div>
+    <nav class="float-left">
+      <ul>
+        <li>
+          <a href="https://www.facebook.com/HongQuan.ITC">
+              {{ __('Tác giả') }}
+          </a>
+        </li>
+        <li>
+          <a href="#">
+              {{ __('HQ Shop') }}
+          </a>
+        </li>
+        <li>
+          <a href="https://hongquanit.com">
+              {{ __('Blog') }}
+          </a>
+        </li>
+      </ul>
+    </nav>
+    <div class="copyright float-right">
+      &copy;
+      <script>
+        document.write(new Date().getFullYear())
+      </script>, made with <i class="material-icons">favorite</i> by
+      <a href="https://www.facebook.com/HongQuan.ITC" target="_blank">Đặng Hồng Quân</a>
+    </div>
     </div>
 </footer>
