+++ sources/dev/resources/views/layouts/navbars/navs/auth-render.blade.php
@@ -11,15 +11,6 @@
     <span class="navbar-toggler-icon icon-bar"></span>
     </button>
     <div class="collapse navbar-collapse justify-content-end">
-      <form class="navbar-form">
-        <div class="input-group no-border">
-        <input type="text" value="" class="form-control" placeholder="Search...">
-        <button type="submit" class="btn btn-white btn-round btn-just-icon">
-          <i class="material-icons">search</i>
-          <div class="ripple-container"></div>
-        </button>
-        </div>
-      </form>
       <ul class="navbar-nav">
         <li class="nav-item">
           <a class="nav-link" href="{{ route('home') }}">
@@ -30,11 +21,9 @@
           </a>
         </li>
         <li class="nav-item">
-            <a class="dropdown-item" href="{{ route('logout') }}" onclick="event.preventDefault();document.getElementById('logout-form').submit();">{{ __('Log out') }}
-                <i class="material-icons">person</i>
-                <p class="d-lg-none d-md-block">
-                {{ __('Account') }}
-                </p>
+            <a class="dropdown-item" href="{{ route('logout') }}" onclick="event.preventDefault();document.getElementById('logout-form').submit();">
+                <i style="margin-right:10px"><img style="width:25px" src="{{ asset('material') }}/img/logout.svg"></i>
+                {{ __('Log out') }}
             </a>
         </li>
       </ul>
