+++ sources/dev/resources/views/layouts/page_templates/auth-render.blade.php
@@ -1,7 +1,7 @@
 <div class="wrapper ">
   @include('layouts.navbars.sidebar')
   <div class="main-panel">
-    @include('layouts.navbars.navs.auth-youtube')
+    @include('layouts.navbars.navs.auth-render')
     @yield('content')
     @include('layouts.footers.auth')
   </div>
