+++ sources/dev/resources/views/welcome.blade.php
@@ -1,10 +1,10 @@
-@extends('layouts.app', ['class' => 'off-canvas-sidebar', 'activePage' => 'home', 'title' => __('Material Dashboard')])
+@extends('layouts.app', ['class' => 'off-canvas-sidebar', 'activePage' => 'home', 'title' => __('HQ Tool')])
 
 @section('content')
 <div class="container" style="height: auto;">
   <div class="row justify-content-center">
       <div class="col-lg-7 col-md-8">
-          <h1 class="text-white text-center">{{ __('Welcome to Material Dashboard FREE Laravel Live Preview.') }}</h1>
+          <h1 class="text-white text-center">{{ __('Chào mừng bạn đến với HQTool. Render video youtube và nhiều thứ khác...') }}</h1>
       </div>
   </div>
 </div>
