+++ sources/dev/webpack.mix.js
@@ -16,12 +16,5 @@ mix.js('resources/js/app.js', 'public/js')
 mix.js('resources/js/pages/youtube.js', 'public/js/pages')
     .sass('resources/sass/youtube.scss', 'public/css/pages');
 mix.sass('resources/sass/material-dashboard.scss', 'public/material/css');
-mix.js('resources/js/jotform/errorNavigation.js', 'public/js/jotform')
-    .sass('resources/sass/jotform/jot.scss', 'public/css');
-mix.js('resources/js/jotform/fileuploader.js', 'public/js/jotform');
-mix.js('resources/js/jotform/imageinfo.js', 'public/js/jotform');
-mix.js('resources/js/jotform/jotform.forms.js', 'public/js/jotform');
-mix.js('resources/js/jotform/prototype.forms.js', 'public/js/jotform');
-mix.js('resources/js/jotform/smoothscroll.min.js', 'public/js/jotform');
